Extract CachedServerTransportParameters in its own file
