+++ quic/client/state/ClientStateMachine.h
@@ -9,6 +9,7 @@
 #pragma once
 
 #include <folly/io/async/AsyncSocketException.h>
+#include <quic/client/handshake/CachedServerTransportParameters.h>
 #include <quic/client/handshake/ClientHandshake.h>
 #include <quic/client/handshake/ClientHandshakeFactory.h>
 #include <quic/congestion_control/QuicCubic.h>
